backend/requirements.txt
# 精简版 requirements.txt - 修复 aioredis 兼容性问题
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis (使用 redis 替代 aioredis)
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Data Sources
pandas==2.1.4
numpy==1.26.3

# HTTP Client
httpx==0.26.0

# Task Queue
celery==5.3.6

# Utilities
python-dotenv==1.0.0
pytz==2023.3
